+++ src/common/chat-participants/powerpages/PowerPagesChatParticipant.ts
@@ -5,26 +5,19 @@
 
 import * as vscode from 'vscode';
 import { createChatParticipant } from '../ChatParticipantUtils';
-import { IPowerPagesChatResult } from './PowerPagesChatParticipantTypes';
+import { IComponentInfo, IPowerPagesChatResult } from './PowerPagesChatParticipantTypes';
 import { ITelemetry } from '../../../client/telemetry/ITelemetry';
 import TelemetryReporter from '@vscode/extension-telemetry';
-import { getIntelligenceEndpoint } from '../../ArtemisService';
 import { sendApiRequest } from '../../copilot/IntelligenceApiService';
 import { PacWrapper } from '../../../client/pac/PacWrapper';
-import { ADX_ENTITYFORM, ADX_ENTITYLIST, COPILOT_UNAVAILABLE, PAC_SUCCESS } from '../../copilot/constants';
-import { createAuthProfileExp, getActiveEditorContent } from '../../Utils';
-import { dataverseAuthentication, intelligenceAPIAuthentication } from '../../AuthenticationProvider';
+import { intelligenceAPIAuthentication } from '../../services/AuthenticationProvider';
 import { ActiveOrgOutput } from '../../../client/pac/PacTypes';
 import { orgChangeErrorEvent, orgChangeEvent } from '../../OrgChangeNotifier';
-import { getEntityName, getFormXml, getEntityColumns } from '../../copilot/dataverseMetadata';
-import { NO_PROMPT_MESSAGE } from './Constants';
-import { MultiStepInput } from '../../../client/power-pages/create/utils/MultiStepInput';
-
-export interface OrgDetails {
-    orgID: string;
-    orgUrl: string;
-}
-
+import { AUTHENTICATION_FAILED_MSG, COPILOT_NOT_AVAILABLE_MSG, NO_PROMPT_MESSAGE, PAC_AUTH_NOT_FOUND, POWERPAGES_CHAT_PARTICIPANT_ID, RESPONSE_AWAITED_MSG } from './PowerPagesChatParticipantConstants';
+import { ORG_DETAILS_KEY, handleOrgChangeSuccess, initializeOrgDetails } from '../../utilities/OrgHandlerUtils';
+import { getComponentInfo, getEndpoint } from './PowerPagesChatParticipantUtils';
+import { checkCopilotAvailability, getActiveEditorContent } from '../../utilities/Utils';
+import { IIntelligenceAPIEndpointInformation } from '../../services/Interfaces';
 export class PowerPagesChatParticipant {
     private static instance: PowerPagesChatParticipant | null = null;
     private chatParticipant: vscode.ChatParticipant;
@@ -33,14 +26,15 @@ export class PowerPagesChatParticipant {
     private readonly _pacWrapper?: PacWrapper;
     private isOrgDetailsInitialized = false;
     private readonly _disposables: vscode.Disposable[] = [];
-    private cachedEndpoint: { intelligenceEndpoint: string, geoName: string } | null = null;
+    private cachedEndpoint: IIntelligenceAPIEndpointInformation| null = null;
 
     private orgID: string | undefined;
-    private orgUrl = '';
+    private orgUrl: string | undefined;
+    private environmentID: string | undefined;
 
     private constructor(context: vscode.ExtensionContext, telemetry: ITelemetry | TelemetryReporter, pacWrapper?: PacWrapper) {
 
-        this.chatParticipant = createChatParticipant('powerpages', this.handler);
+        this.chatParticipant = createChatParticipant(POWERPAGES_CHAT_PARTICIPANT_ID, this.handler);
 
         //TODO: Check the icon image
         this.chatParticipant.iconPath = vscode.Uri.joinPath(context.extensionUri, 'src', 'common', 'chat-participants', 'powerpages', 'assets', 'copilot.png');
@@ -56,17 +50,8 @@ export class PowerPagesChatParticipant {
         }));
 
         this._disposables.push(orgChangeErrorEvent(async () => {
-            await createAuthProfileExp(this._pacWrapper);
+            this.extensionContext.globalState.update(ORG_DETAILS_KEY, { orgID: undefined, orgUrl: undefined });
         }));
-
-        vscode.commands.registerCommand('create-site-inputs', async (siteName: string, isCreateSiteInputsReceived) => {
-            if (!isCreateSiteInputsReceived) {
-                const siteCreateInputs = await this.getSiteCreationInputs(siteName);
-                if(siteCreateInputs){
-                    isCreateSiteInputsReceived = true;
-                }
-            }
-        });
 
     }
 
@@ -86,17 +71,16 @@ export class PowerPagesChatParticipant {
         request: vscode.ChatRequest,
         _context: vscode.ChatContext,
         stream: vscode.ChatResponseStream,
-        _token: vscode.CancellationToken
+        //_token: vscode.CancellationToken
     ): Promise<IPowerPagesChatResult> => {
         // Handle chat requests here
 
-        stream.progress('Working on it...')
+        stream.progress(RESPONSE_AWAITED_MSG)
 
-        await this.intializeOrgDetails();
+        await this.initializeOrgDetails();
 
-        if (!this.orgID) {
-            // TODO: Auth Create Experience using button
-            await createAuthProfileExp(this._pacWrapper);
+        if (!this.orgID || !this.environmentID) {
+            stream.markdown(PAC_AUTH_NOT_FOUND);
             return {
                 metadata: {
                     command: ''
@@ -107,8 +91,7 @@ export class PowerPagesChatParticipant {
         const intelligenceApiAuthResponse = await intelligenceAPIAuthentication(this.telemetry, '', this.orgID, true);
 
         if (!intelligenceApiAuthResponse) {
-
-            //TODO: Handle auth error and provide a way to re-authenticate
+            stream.markdown(AUTHENTICATION_FAILED_MSG);
 
             return {
                 metadata: {
@@ -119,34 +102,22 @@ export class PowerPagesChatParticipant {
 
         const intelligenceApiToken = intelligenceApiAuthResponse.accessToken;
 
-        const { intelligenceEndpoint, geoName } = await this.getEndpoint(this.orgID, this.telemetry);
+        const intelligenceAPIEndpointInfo = await getEndpoint(this.orgID, this.environmentID, this.telemetry, this.cachedEndpoint);
 
-        const endpointAvailabilityResult = this.handleEndpointAvailability(intelligenceEndpoint, geoName);
+        const copilotAvailabilityStatus = checkCopilotAvailability(intelligenceAPIEndpointInfo.intelligenceEndpoint, this.orgID, this.telemetry, '');
 
-        if (endpointAvailabilityResult !== '') {
-            return endpointAvailabilityResult;
+        if (!copilotAvailabilityStatus) {
+            stream.markdown(COPILOT_NOT_AVAILABLE_MSG)
+
+            return {
+                metadata: {
+                    command: ''
+                }
+            };
         }
 
         if (request.command) {
             //TODO: Handle command scenarios
-            if(request.command == 'create-site')
-            {
-                stream.progress('Generating a new Power Pages site...');
-
-                stream.markdown('Below is the markdown content for the new Power Pages site. You can copy this content and paste it in the markdown file to create a new Power Pages site.');
-
-
-
-                stream.button({
-                    command: 'create-site-inputs',
-                    title: 'Create Site',
-                    tooltip: 'Create a new Power Pages site',
-                    arguments: ['siteName', false],
-                })
-
-                 // register command
-
-            }
 
         } else {
 
@@ -164,29 +135,11 @@ export class PowerPagesChatParticipant {
                 };
             }
 
-            const { activeFileParams } = getActiveEditorContent();
+            const {activeFileParams} = getActiveEditorContent();
 
-            let metadataInfo = { entityName: '', formName: '' };
-            let componentInfo: string[] = [];
+            const {componentInfo, entityName}: IComponentInfo = await getComponentInfo(this.telemetry, this.orgUrl, activeFileParams);
 
-
-            if (activeFileParams.dataverseEntity == ADX_ENTITYFORM || activeFileParams.dataverseEntity == ADX_ENTITYLIST) {
-                metadataInfo = await getEntityName(this.telemetry, '', activeFileParams.dataverseEntity);
-
-                const dataverseToken = (await dataverseAuthentication(this.telemetry, this.orgUrl, true)).accessToken;
-
-                if (activeFileParams.dataverseEntity == ADX_ENTITYFORM) {
-                    const formColumns = await getFormXml(metadataInfo.entityName, metadataInfo.formName, this.orgUrl, dataverseToken, this.telemetry, 'sessionID');
-                    componentInfo = formColumns;
-                } else {
-                    const entityColumns = await getEntityColumns(metadataInfo.entityName, this.orgUrl, dataverseToken, this.telemetry, 'sessionID');
-                    componentInfo = entityColumns;
-                }
-
-            }
-
-            // export async function sendApiRequest(userPrompt: UserPrompt[], activeFileParams: IActiveFileParams, orgID: string, apiToken: string, sessionID: string, entityName: string, entityColumns: string[], telemetry: ITelemetry, aibEndpoint: string | null, geoName: string | null) {}
-            const llmResponse = await sendApiRequest([{ displayText: userPrompt, code: '' }], activeFileParams, this.orgID, intelligenceApiToken, '', '', componentInfo, this.telemetry, intelligenceEndpoint, geoName);
+            const llmResponse = await sendApiRequest([{ displayText: userPrompt, code: '' }], activeFileParams, this.orgID, intelligenceApiToken, '', entityName, componentInfo, this.telemetry, intelligenceAPIEndpointInfo.intelligenceEndpoint, intelligenceAPIEndpointInfo.geoName, intelligenceAPIEndpointInfo.crossGeoDataMovementEnabledPPACFlag);
 
             llmResponse.forEach((response: { displayText: string | vscode.MarkdownString; code: string; }) => {
                 if (response.displayText) {
@@ -200,8 +153,6 @@ export class PowerPagesChatParticipant {
 
         }
 
-        console.log(_token)
-
         return {
             metadata: {
                 command: ''
@@ -210,145 +161,17 @@ export class PowerPagesChatParticipant {
 
     };
 
-    private async intializeOrgDetails(): Promise<void> {
-
-        if (this.isOrgDetailsInitialized) {
-            return;
-        }
-
-        this.isOrgDetailsInitialized = true;
-
-        const orgDetails: OrgDetails | undefined = this.extensionContext.globalState.get('orgDetails');
-
-        if (orgDetails) {
-            this.orgID = orgDetails.orgID;
-            this.orgUrl = orgDetails.orgUrl;
-        } else {
-            if (this._pacWrapper) {
-                const pacActiveOrg = await this._pacWrapper.activeOrg();
-                if (pacActiveOrg && pacActiveOrg.Status === PAC_SUCCESS) {
-                    this.handleOrgChangeSuccess(pacActiveOrg.Results);
-                } else {
-                    await createAuthProfileExp(this._pacWrapper);
-                }
-            }
-        }
+    private async initializeOrgDetails(): Promise<void> {
+        const { orgID, orgUrl, environmentID } = await initializeOrgDetails(this.isOrgDetailsInitialized, this.extensionContext, this._pacWrapper);
+        this.orgID = orgID;
+        this.orgUrl = orgUrl;
+        this.environmentID = environmentID;
     }
 
-    private async handleOrgChangeSuccess(orgDetails: ActiveOrgOutput) {
-        this.orgID = orgDetails.OrgId;
-        this.orgUrl = orgDetails.OrgUrl
-
-        this.extensionContext.globalState.update('orgDetails', { orgID: this.orgID, orgUrl: this.orgUrl });
-
-        //TODO: Handle AIB GEOs
-
-        this.cachedEndpoint = null;
+    private async handleOrgChangeSuccess(orgDetails: ActiveOrgOutput): Promise<void> {
+        const { orgID, orgUrl, environmentID } = handleOrgChangeSuccess(orgDetails, this.extensionContext);
+        this.orgID = orgID;
+        this.orgUrl = orgUrl;
+        this.environmentID = environmentID;
     }
-
-    async getEndpoint(orgID: string, telemetry: ITelemetry) {
-        if (!this.cachedEndpoint) {
-            this.cachedEndpoint = await getIntelligenceEndpoint(orgID, telemetry, '') as { intelligenceEndpoint: string; geoName: string };
-        }
-        return this.cachedEndpoint;
-    }
-
-    handleEndpointAvailability(intelligenceEndpoint: string, geoName: string) {
-        if (!intelligenceEndpoint || !geoName) {
-            return {
-                metadata: {
-                    command: ''
-                }
-            };
-        } else if (intelligenceEndpoint === COPILOT_UNAVAILABLE) {
-
-            //TODO: Handle Copilot Unavailable scenario with response to user
-            return {
-                metadata: {
-                    command: ''
-                }
-            };
-        }
-
-        //TODO: Handle ECS unavailable scenario
-
-        return '' //TODO return type
-    }
-
-    async getSiteCreationInputs(siteName: string) {
-        const envNames: vscode.QuickPickItem[] = [
-            { label: 'EnvONe' },
-            { label: 'EnvTwo' },
-            { label: 'EnvThree' }
-        ];
-
-        const title = vscode.l10n.t("New Power Pages Site");
-
-        interface ISiteInputState {
-            siteName: string;
-            envName: string;
-            domainName: string;
-            title: string;
-            step: number;
-            totalSteps: number;
-        }
-
-        async function collectInputs() {
-            const state = {} as Partial<ISiteInputState>;
-            await MultiStepInput.run((input) => selectEnvName(input, state));
-            return state as ISiteInputState;
-        }
-
-        async function selectEnvName(
-            input: MultiStepInput,
-            state: Partial<ISiteInputState>
-        ) {
-            const pick = await input.showQuickPick({
-                title,
-                step: 1,
-                totalSteps: 3,
-                placeholder: vscode.l10n.t("Choose Environment"),
-                items: envNames,
-                activeItem:
-                    typeof state.envName !== "string"
-                        ? state.envName
-                        : undefined,
-            });
-            state.envName = pick.label;
-            return (input: MultiStepInput) => inputSiteName(input, state);
-        }
-
-        async function inputSiteName(
-            input: MultiStepInput,
-            state: Partial<ISiteInputState>
-        ) {
-            state.siteName = await input.showInputBox({
-                title,
-                step: 2,
-                totalSteps: 3,
-                value: state.siteName || siteName,
-                placeholder: vscode.l10n.t("Enter Site Name"),
-                validate: async (value) => (value ? undefined : vscode.l10n.t("Site Name is required")),
-            });
-            return (input: MultiStepInput) => inputDomainName(input, state);
-        }
-
-        async function inputDomainName(
-            input: MultiStepInput,
-            state: Partial<ISiteInputState>
-        ) {
-            state.domainName = await input.showInputBox({
-                title,
-                step: 3,
-                totalSteps: 3,
-                value: state.domainName || "",
-                placeholder: vscode.l10n.t("Enter Domain Name"),
-                validate: async (value) => (value ? undefined : vscode.l10n.t("Domain Name is required")),
-            });
-        }
-
-        const siteInputState = await collectInputs();
-        return siteInputState;
-    }
-
 }