+++ src/common/services/AuthenticationProvider.ts
@@ -4,19 +4,11 @@
  */
 
 import * as vscode from "vscode";
-import {
-    INTELLIGENCE_SCOPE_DEFAULT,
-    PROVIDER_ID,
-    SCOPE_OPTION_CONTACTS_READ,
-    SCOPE_OPTION_DEFAULT,
-    SCOPE_OPTION_OFFLINE_ACCESS,
-    SCOPE_OPTION_USERS_READ_BASIC_ALL,
-} from "../web/client/common/constants";
-import { showErrorDialog } from "../web/client/common/errorHandler";
-import { ITelemetry } from "../client/telemetry/ITelemetry";
-import { sendTelemetryEvent } from "./copilot/telemetry/copilotTelemetry";
-import { CopilotLoginFailureEvent, CopilotLoginSuccessEvent } from "./copilot/telemetry/telemetryConstants";
-import { getUserAgent } from "./Utils";
+import { showErrorDialog } from "../../web/client/common/errorHandler";
+import { ITelemetry } from "../../client/telemetry/ITelemetry";
+import { sendTelemetryEvent } from "../copilot/telemetry/copilotTelemetry";
+import { CopilotLoginFailureEvent, CopilotLoginSuccessEvent } from "../copilot/telemetry/telemetryConstants";
+import { getUserAgent } from "../utilities/Utils";
 import {
     VSCODE_EXTENSION_DATAVERSE_AUTHENTICATION_COMPLETED,
     VSCODE_EXTENSION_DATAVERSE_AUTHENTICATION_FAILED,
@@ -24,9 +16,12 @@ import {
     VSCODE_EXTENSION_NPS_AUTHENTICATION_FAILED,
     VSCODE_EXTENSION_NPS_AUTHENTICATION_STARTED,
     VSCODE_EXTENSION_GRAPH_CLIENT_AUTHENTICATION_FAILED,
-    VSCODE_EXTENSION_GRAPH_CLIENT_AUTHENTICATION_COMPLETED
+    VSCODE_EXTENSION_GRAPH_CLIENT_AUTHENTICATION_COMPLETED,
+    VSCODE_EXTENSION_BAP_SERVICE_AUTHENTICATION_COMPLETED,
+    VSCODE_EXTENSION_BAP_SERVICE_AUTHENTICATION_FAILED
 } from "./TelemetryConstants";
-import { ERRORS } from "./ErrorConstants";
+import { ERRORS } from "../ErrorConstants";
+import { BAP_SERVICE_SCOPE_DEFAULT, INTELLIGENCE_SCOPE_DEFAULT, PROVIDER_ID, SCOPE_OPTION_CONTACTS_READ, SCOPE_OPTION_DEFAULT, SCOPE_OPTION_OFFLINE_ACCESS, SCOPE_OPTION_USERS_READ_BASIC_ALL } from "./Constants";
 
 
 export function getCommonHeadersForDataverse(
@@ -55,8 +50,6 @@ export function getCommonHeaders(
             ? "application/octet-stream"
             : "application/json; charset=utf-8",
         accept: "application/json",
-        "OData-MaxVersion": "4.0",
-        "OData-Version": "4.0"
     };
 }
 
@@ -84,10 +77,9 @@ export async function intelligenceAPIAuthentication(telemetry: ITelemetry, sessi
             sendTelemetryEvent(telemetry, { eventName: CopilotLoginSuccessEvent, copilotSessionId: sessionID, orgId: orgId });
         }
     } catch (error) {
-        const authError = (error as Error)
         showErrorDialog(vscode.l10n.t("Authorization Failed. Please run again to authorize it"),
             vscode.l10n.t("There was a permissions problem with the server"));
-        sendTelemetryEvent(telemetry, { eventName: CopilotLoginFailureEvent, copilotSessionId: sessionID, orgId: orgId, error: authError });
+        sendTelemetryEvent(telemetry, { eventName: CopilotLoginFailureEvent, copilotSessionId: sessionID, orgId: orgId, errorMsg: (error as Error).message });
     }
     return { accessToken, user, userId };
 }
@@ -145,7 +137,7 @@ export async function dataverseAuthentication(
         sendTelemetryEvent(
             telemetry, {
             eventName: VSCODE_EXTENSION_DATAVERSE_AUTHENTICATION_FAILED,
-            error: error as Error
+            errorMsg: (error as Error).message
         }
         );
     }
@@ -185,7 +177,7 @@ export async function npsAuthentication(
             telemetry,
             {
                 eventName: VSCODE_EXTENSION_NPS_AUTHENTICATION_FAILED,
-                error: error as Error
+                errorMsg: (error as Error).message
             }
         );
     }
@@ -241,7 +233,56 @@ export async function graphClientAuthentication(
             vscode.l10n.t("There was a permissions problem with the server")
         );
         sendTelemetryEvent(telemetry,
-            { eventName: VSCODE_EXTENSION_GRAPH_CLIENT_AUTHENTICATION_FAILED, error: error as Error }
+            { eventName: VSCODE_EXTENSION_GRAPH_CLIENT_AUTHENTICATION_FAILED, errorMsg: (error as Error).message }
+        )
+    }
+
+    return accessToken;
+}
+
+export async function bapServiceAuthentication(
+    telemetry: ITelemetry,
+    firstTimeAuth = false
+): Promise<string> {
+    let accessToken = "";
+    try {
+        let session = await vscode.authentication.getSession(
+            PROVIDER_ID,
+            [BAP_SERVICE_SCOPE_DEFAULT],
+            { silent: true }
+        );
+
+        if (!session) {
+            session = await vscode.authentication.getSession(
+                PROVIDER_ID,
+                [BAP_SERVICE_SCOPE_DEFAULT],
+                { createIfNone: true }
+            );
+        }
+
+        accessToken = session?.accessToken ?? "";
+        if (!accessToken) {
+            throw new Error(ERRORS.NO_ACCESS_TOKEN);
+        }
+
+        if (firstTimeAuth) {
+            sendTelemetryEvent(telemetry, {
+                eventName: VSCODE_EXTENSION_BAP_SERVICE_AUTHENTICATION_COMPLETED,
+                userId:
+                    session?.account.id.split("/").pop() ??
+                    session?.account.id ??
+                    "",
+            });
+        }
+    } catch (error) {
+        showErrorDialog(
+            vscode.l10n.t(
+                "Authorization Failed. Please run again to authorize it"
+            ),
+            vscode.l10n.t("There was a permissions problem with the server")
+        );
+        sendTelemetryEvent(telemetry,
+            { eventName: VSCODE_EXTENSION_BAP_SERVICE_AUTHENTICATION_FAILED, errorMsg: (error as Error).message }
         )
     }