+++ src/common/utilities/Utils.ts
@@ -3,13 +3,16 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-
 import * as vscode from "vscode";
-import { EXTENSION_ID, EXTENSION_NAME, SETTINGS_EXPERIMENTAL_STORE_NAME } from "../client/constants";
-import { CUSTOM_TELEMETRY_FOR_POWER_PAGES_SETTING_NAME } from "./OneDSLoggerTelemetry/telemetryConstants";
-import { PacWrapper } from "../client/pac/PacWrapper";
-import { AUTH_CREATE_FAILED, AUTH_CREATE_MESSAGE, DataverseEntityNameMap, EntityFieldMap, FieldTypeMap, PAC_SUCCESS } from "./copilot/constants";
-import { IActiveFileData, IActiveFileParams } from "./copilot/model";
+import { EXTENSION_ID, EXTENSION_NAME, SETTINGS_EXPERIMENTAL_STORE_NAME } from "../../client/constants";
+import { CUSTOM_TELEMETRY_FOR_POWER_PAGES_SETTING_NAME } from "../OneDSLoggerTelemetry/telemetryConstants";
+import { PacWrapper } from "../../client/pac/PacWrapper";
+import { AUTH_CREATE_FAILED, AUTH_CREATE_MESSAGE, COPILOT_UNAVAILABLE, DataverseEntityNameMap, EntityFieldMap, FieldTypeMap, PAC_SUCCESS } from "../copilot/constants";
+import { IActiveFileData, IActiveFileParams } from "../copilot/model";
+import { ITelemetry } from "../../client/telemetry/ITelemetry";
+import { sendTelemetryEvent } from "../copilot/telemetry/copilotTelemetry";
+import { getDisabledOrgList, getDisabledTenantList } from "../copilot/utils/copilotUtil";
+import { CopilotNotAvailable, CopilotNotAvailableECSConfig } from "../copilot/telemetry/telemetryConstants";
 
 export function getSelectedCode(editor: vscode.TextEditor): string {
     if (!editor) {
@@ -92,7 +95,8 @@ export function showConnectedOrgMessage(environmentName: string, orgUrl: string)
 export async function showInputBoxAndGetOrgUrl() {
     return vscode.window.showInputBox({
         placeHolder: vscode.l10n.t("Enter the environment URL"),
-        prompt: vscode.l10n.t("Active auth profile is not found or has expired. To create a new auth profile, enter the environment URL.")
+        prompt: vscode.l10n.t("Active auth profile is not found or has expired. To create a new auth profile, enter the environment URL."),
+        ignoreFocusOut: true
     });
 }
 
@@ -120,7 +124,7 @@ export function openWalkthrough(extensionUri: vscode.Uri) {
     vscode.commands.executeCommand("markdown.showPreview", walkthroughUri);
 }
 
-export function isCustomTelemetryEnabled():boolean {
+export function isCustomTelemetryEnabled(): boolean {
     const isCustomTelemetryEnabled = vscode.workspace
         .getConfiguration(SETTINGS_EXPERIMENTAL_STORE_NAME)
         .get(CUSTOM_TELEMETRY_FOR_POWER_PAGES_SETTING_NAME);
@@ -133,7 +137,7 @@ export function getUserAgent(): string {
     return userAgent
         .replace("{product}", EXTENSION_NAME)
         .replace("{product-version}", getExtensionVersion())
-        .replace("{comment}", "(" + getExtensionType()+'; )');
+        .replace("{comment}", "(" + getExtensionType() + '; )');
 }
 
 export async function createAuthProfileExp(pacWrapper: PacWrapper | undefined) {
@@ -142,7 +146,7 @@ export async function createAuthProfileExp(pacWrapper: PacWrapper | undefined) {
         return;
     }
 
-    if(!pacWrapper){
+    if (!pacWrapper) {
         vscode.window.showErrorMessage(AUTH_CREATE_FAILED);
         return;
     }
@@ -178,4 +182,26 @@ export function getActiveEditorContent(): IActiveFileData {
     }
 
     return activeFileData;
+}
+
+export function checkCopilotAvailability(
+    aibEndpoint: string | null,
+    orgID: string,
+    telemetry: ITelemetry,
+    sessionID: string,
+    tenantId?: string | undefined,
+): boolean {
+
+    if(!aibEndpoint) {
+        return false;
+    }
+    else if (aibEndpoint === COPILOT_UNAVAILABLE ) {
+        sendTelemetryEvent(telemetry, { eventName: CopilotNotAvailable, copilotSessionId: sessionID, orgId: orgID });
+        return false;
+    } else if (getDisabledOrgList()?.includes(orgID) || getDisabledTenantList()?.includes(tenantId ?? "")) { // Tenant ID not available in desktop
+        sendTelemetryEvent(telemetry, { eventName: CopilotNotAvailableECSConfig, copilotSessionId: sessionID, orgId: orgID });
+        return false;
+    } else {
+        return true;
+    }
 }