+++ src/web/client/webViews/QuickPickProvider.ts
@@ -12,7 +12,6 @@ import { getFileEntityId, getFileEntityName, getFileRootWebPageId } from "../uti
 interface IQuickPickItem extends vscode.QuickPickItem {
     label: string;
     id?: string;
-    iconPath?: string | vscode.Uri | { light: string | vscode.Uri; dark: string | vscode.Uri } | vscode.ThemeIcon;
 }
 
 export class QuickPickProvider {
@@ -23,14 +22,18 @@ export class QuickPickProvider {
     }
 
     public refresh() {
-        if (vscode.window.activeTextEditor) {
-            const fileFsPath = vscode.window.activeTextEditor.document.uri.fsPath;
-            const entityInfo: IEntityInfo = {
-                entityId: getFileEntityId(fileFsPath),
-                entityName: getFileEntityName(fileFsPath),
-                rootWebPageId: getFileRootWebPageId(fileFsPath),
-            };
-            this.updateQuickPickItems(entityInfo);
+        const tabGroup = vscode.window.tabGroups;
+        if (tabGroup.activeTabGroup && tabGroup.activeTabGroup.activeTab) {
+            const tab = tabGroup.activeTabGroup.activeTab;
+            if (tab.input instanceof vscode.TabInputCustom || tab.input instanceof vscode.TabInputText) {
+                const fileFsPath = tab.input.uri.fsPath;
+                const entityInfo: IEntityInfo = {
+                    entityId: getFileEntityId(fileFsPath),
+                    entityName: getFileEntityName(fileFsPath),
+                    rootWebPageId: getFileRootWebPageId(fileFsPath),
+                };
+                this.updateQuickPickItems(entityInfo);
+            }
         }
     }